+++ R/hx_attrs.R
@@ -6,22 +6,42 @@
 hx_attrs <- function(
   get = NULL,
   post = NULL,
+  put = NULL,
+  patch = NULL,
+  delete = NULL,
   target = NULL,
   swap = NULL,
   trigger = NULL,
   indicator = NULL,
   swap_oob = NULL,
-  confirm = NULL
+  confirm = NULL,
+  params = NULL,
+  include = NULL,
+  push_url = NULL,
+  select = NULL,
+  vals = NULL,
+  encoding = NULL,
+  headers = NULL
 ) {
   attrs <- list(
     `hx-get` = get,
     `hx-post` = post,
+    `hx-put` = put,
+    `hx-patch` = patch,
+    `hx-delete` = delete,
     `hx-target` = target,
     `hx-swap` = swap,
     `hx-trigger` = trigger,
     `hx-indicator` = indicator,
     `hx-swap-oob` = swap_oob,
-    `hx-confirm` = confirm
+    `hx-confirm` = confirm,
+    `hx-params` = params,
+    `hx-include` = include,
+    `hx-push-url` = push_url,
+    `hx-select` = select,
+    `hx-vals` = vals,
+    `hx-encoding` = encoding,
+    `hx-headers` = headers
   )
   attrs[!vapply(attrs, is.null, logical(1))]
 }